sched-ext: introduce the concept of performance profiles
Add an extra combo box to select the "performance profile" for the
currently selected sched-ext scheduler. The selected profile will be
automatically converted into the appropriate flags for the chosen
scheduler.

Currently, only scx_bpfland and scx_lavd are supported, so the
additional combo box will be visible only when one of these schedulers
is selected.

Tested-by: Peter Jung <ptr1337@cachyos.org>
Signed-off-by: Andrea Righi <andrea.righi@linux.dev>