sched_ext: drop -c 0 from the gaming preset with scx_bpfland
Running bpfland with `-c 0` can lead to stuttery behavior if multiple
worklodas are running in a system at the same time.

Since this option offers minimal benefits for gaming and is only
effective when a *single* application is running, it's just safer to
avoid using it in the gaming profile.

Signed-off-by: Andrea Righi <andrea.righi@linux.dev>